+++ programs/registry/src/lib.rs
@@ -360,7 +360,7 @@ pub mod light_registry {
         ctx: Context<'_, '_, '_, 'info, NullifyLeaves<'info>>,
         bump: u8,
         change_log_indices: Vec<u64>,
-        leaves_queue_indices: Vec<u16>,
+        leaves_queue_indices: Vec<u64>,
         indices: Vec<u64>,
         proofs: Vec<Vec<[u8; 32]>>,
     ) -> Result<()> {
